renamed components extension .js to .jsx, functional
