+++ src/resources/components/Account.jsx
@@ -1,18 +1,20 @@
 import React, { useState } from "react";
-import { Menu } from "./Menu.js";
-import { NavBar } from "./NavBar.js";
+import { Menu } from "./Menu";
+import { NavBar } from "./NavBar";
 
-export function ModelX() {
+export function Account() {
   const [menuCoverClass, setMenuCoverClass] = useState(
     "menuCoverPage displayNone"
   );
   return (
-    <div className="ModelX">
+    <div className="Shop">
       <header>
+        {/* This navbar must change for this component */}
         <NavBar setMenuCoverClasss={setMenuCoverClass} />
       </header>
       <main>
-        Model X
+        Account
+        {/*  This menu must change for this component */}
         <Menu
           menuCoverClasss={menuCoverClass}
           setMenuCoverClasss={setMenuCoverClass}
